requirements.txt
playwright
openai==0.28.1
python-dotenv
asyncio